updateControllerOk

+++ src/controllers/ImageController.js
@@ -1,5 +1,6 @@
 const knex = require("../database/knex");
 const DiskStorage = require("../providers/DiskStorage");
+const AppError = require("../utils/AppError");
 
 class ImageController {
   async update(request, response) {
@@ -9,6 +10,13 @@ class ImageController {
     const diskStorage = new DiskStorage();
 
     const dish = await knex("dishes").where({ id: dish_id }).first();
+
+    if (!dish) {
+      throw new AppError(
+        "O prato selecionado para fazer o upload da imagem não existe. Tente novamente.",
+        401
+      );
+    }
 
     if (dish.image) {
       await diskStorage.deleteFile(dish.image);
